Christoph Biedl <sourceforge.bnwi@manchmal.in-ulm.de> adds the feature of offset and size command-line options.
In the quote below are his stated motivations:

Hello,

the patch below implements a new option -o for vbladed, denoting the
number of sectors that should be skipped at the beginning of the
exported file. Default is 0 (no offset, export from beginning, current
behaviour).

Rationale:

If the exported files are actually raw block devices on the target,
something that might happen when using logical volumes for exports,
several tools running on the target that assume block devices are
for exclusive local access only might create confusion or even
havoc. Most prominently:

* os-prober of grub (might be disabled, though)
 Might mount any block device found that contain a file system.
* (Linux) logical volume management
 Will report any physical volume if the export is used as such.
* (Linux) blkid
 Reports any block device (minor impact, mostly information leakage)

The overall problem is applications are not aware some block devices
might be used from a remote client while there's no locking and that
might be hard to implement

Solutions tried but not considered helpful:

* Filter out exported block devices
 Problems: The tools need the ability to filter; new tools might come
 into existance, requiring new action. Filter languages might be hard
 to use (Linux LVM is quite a nightmare here).
* Optionally open block devices with O_EXCL
 Had no effect. Still exposes block devices where no vbladed is running
 yet.

So the goal was to hide the actual content from these tools. One
approach was to obscure the data using XOR, however it seemed easier
to sacrifice one or more sectors (8 recommended for 4k sector hard
drives) for a simple workaround.

This is based on vblade-20, tests on Debian Linux included a simple
write of each sector with a unique ID and later re-read in order to
detect any re-ordering bugs, however everything worked as expected.

   Christoph