CHS misalignment warning
WinAoE has some code in the GettingsSize state that truncates a disk to
CHS geometry. Prior to Vista, Windows enforced CHS alignment for
partition boundaries, so this was not a problem.
However, if you installed a newer OS (one using 1MB boundaries) then
moved it to AoE storage, truncating at a partition boundary could cause
sectors to be missing under WinAoE, corrupting your data. Windows
probably never actually relied on this behaviour, since it was enforcing
alignment itself.
This commit adds a warning regarding this to the man page.